football_lms/backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework @ git+https://github.com/encode/django-rest-framework.git@dbac145638758413b966c3418fa5f3f651e3e02a
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==1.7.1
python-dotenv==1.0.1
setuptools==75.6.0
sqlparse==0.5.2
